requirements.txt
diffusers==0.29.2
transformers==4.41.3
accelerate==0.31.0
Pillow==10.4.0
tqdm==4.66.4
fastapi==0.111.0
uvicorn[standard]==0.30.1
safetensors==0.4.3
numpy==1.26.4
